This return type was wrong. Song progress should now work again.

+++ iTunes.h
@@ -145,7 +145,7 @@ typedef enum {
 @property BOOL fullScreen;  // are visuals displayed using the entire screen?
 @property (copy, readonly) NSString *name;  // the name of the application
 @property BOOL mute;  // has the sound output been muted?
-@property NSInteger playerPosition;  // the player’s position within the currently playing track in seconds.
+@property double playerPosition;  // the player’s position within the currently playing track in seconds.
 @property (readonly) iTunesEPlS playerState;  // is iTunes stopped, paused, or playing?
 @property (copy, readonly) SBObject *selection;  // the selection visible to the user
 @property NSInteger soundVolume;  // the sound output volume (0 = minimum, 100 = maximum)
